Feature Selection for Monotonic Classification

Selects the N most relevant features from ordinal and monotonic data, based on mRMR criterion.

fselector(traindata, trainlabels, k, beta, nselected)

Arguments

traindata
Training data of numeric type without labels.

trainlabels
A vector of numeric tags for each instance of training data.

k
positive constant for logit function. If large, fuzzy ordinal set is understood as slightly larger. If small, is understood as significantly larger.

beta
Regulation param for relative importance of MI between features and decision.

nselected
Number of features to select.

Value

nselected most important features.

Examples

dattrain<-read.table("train_balance-scale.0", sep=" ")

#> Warning: cannot open file 'train_balance-scale.0': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

trainlabels<-dattrain[,ncol(dattrain)]

#> Error in eval(expr, envir, enclos): object 'dattrain' not found

traindata=dattrain[,-ncol(dattrain)]

#> Error in eval(expr, envir, enclos): object 'dattrain' not found

selected<-fselector(traindata,trainlabels,2,2,2)

#> Error in fselector(traindata, trainlabels, 2, 2, 2): object 'trainlabels' not found